% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaccard.R
\name{jaccard}
\alias{jaccard}
\title{Jaccard similarity matrix.}
\usage{
jaccard(x, y)
}
\arguments{
\item{x, y}{-- vectors of cluster labels}
}
\value{
a matrix of Jaccard similarity between clusters in two partitions.
}
\description{
This function calculates Jaccard similarity matrix between two partitions.
}
\examples{
x=c(1,2,3)
y=c(3,2,1)
jaccard(x,y)
}
