% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim1}
\alias{sim1}
\title{Simulated toy data}
\format{A matrix with 5000 rows and 2 variables}
\usage{
sim1
}
\description{
A dataset containing 5000 samples and 2 features.
}
\keyword{datasets}
