% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vis_pollen}
\alias{vis_pollen}
\title{Single cell gene visualization data from Pollen's paper}
\format{A list containing two components
\describe{
  \item{vis}{visualization coordinates of cells}
  \item{ref}{true labels of cells}
}}
\source{
\url{https://www.nature.com/articles/nbt.2967}
}
\usage{
vis_pollen
}
\description{
A dataset containing the visualization coordinates and the true cluster labels of 301 cells.
}
\keyword{datasets}
