% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wassDist.R
\name{wassDist}
\alias{wassDist}
\title{Wasserstein distance between two partitions.}
\usage{
wassDist(x, y)
}
\arguments{
\item{x, y}{-- vectors of cluster labels}
}
\value{
a distance between 0 and 1.
}
\description{
This function calculates Wasserstein distance between two partitions.
}
\examples{
x=c(1,2,3)
y=c(3,2,1)
wassDist(x,y)
}
