% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{YAN}
\alias{YAN}
\title{Single cell gene data from Yan's paper}
\format{A matrix with 124 rows and 3840 variables}
\source{
\url{https://www.nature.com/articles/nsmb.2660}
}
\usage{
YAN
}
\description{
A dataset containing 124 cells with their 3840 genes.
}
\keyword{datasets}
